telegram_bot/requirements.txt
python-telegram-bot==20.6
PyMySQL==1.1.1
pycryptodome
python-dotenv
